Sort sources by distance to controller
The logic here is that we want creeps to start with the closer sources
(that way the first creep spawned will be able to fill the controller
up faster). To save CPU it looks for the controller rather than doing a
check for all spawns (this is taking advantage of the fact that the
quorum code prioritizes putting the spawn close to the controller).