Add ability to set default energy for roles
Before this PR the system would scale creeps up to the size of the room
if an energy amount was provided, and if an energy option was provided
that was greater than the capacity of the room (due to lost extensions
or buggy code calling it) it would fail to spawn and stall out the
system.

This update allows a new property to be set on each role which defines
the default value if one isn’t provided. It also automatically reduces
that amount when pulling the creep from the queue if the room
energyCapacityAvailable is lower than it.

To do this the PR also adds the room as a parameter to the getBuild
function, which will allow other room based logic to take place when
deciding the build for a creep role