Rebuild destroyed rooms using neighbors
This adds a new `recover` mode to the `expand` program which operates
in a more limited way (as the room is already claimed and running its
own programs), sending out upgraders and builders to the room in order
to get it rebuilt.

It also better handles failed cities (which have completely timed out)
by removing them from the `city` list.