Reduce upgraders on recovering rooms
Reduce upgraders if a room is being assisted from its neighbor or is
building itself up from one level to another (as in, its current
practical room level is lower than the room level itself). This way the
energy in the room will be focused on building.

If the room is in danger of downgrading upgraders will still be sent in
by the expansion program and spawned by the room’s emergency upgrader
protocols.