If room candidate list is empty refill it.
When there is only a single room and the `expand` program runs it may
define the `candidateList` as an empty array. This happens because the
`expand` program launches immediately but there are no rooms capable of
expanding until the first one hits `PRL4`.  This PR fixes this by
checking for empty arrays and attempting to rebuild the list should
they occur.