Configure prettier for formatting
Standard is good, but prettier is even better. Prettier takes
auto-formatting to the next level.

This change leverages eslint to keep using standard for all the linting
that checks for code errors, but switches to using prettier for code
formatting.

Developers will need to install editor plugins for Eslint and Prettier
and disable any plugin that they previously had for Standard. It is also
recommended for developers to set the plugins to auto-fix/auto-format
for maximum awesomeness.

Prettier: https://prettier.io/