Raise STRUCTURE_STORAGE in building priority
As the storage is a key structure for resource based operations we
raise the priority so it's built before additional extensions on
level up.
This might have a negative impact on restoring rooms but will reduce some inefficencies regarding or haulers, which highly depend on a structure with a lot of capacity. Additionally this increases the speed a colony
advances to a truely self sufficient city, resulting in faster expansion.