Add system to recover from a drained bucket
If the bucket drops below `BUCKET_EMERGENCY` then the system will
attempt to bring the bucket back up by not running anything until the
system gets to `BUCKET_CEILING`. It will only attempt this once every
`BUCKET_BUILD_LIMIT` (currently 15000) ticks.

The bucket should never get this low, but occasionally when it does
(due to errors such as like launching a billion mines) the kernel can
oscillate between not running much at all (due to being below the
limit) and then attempting to start back up (with empty caches) and
using enough CPU to drop below the limit again. In situations where
this really is temporary then this system will do great, but if it is a
real problem this won’t do much.