Upgrade consensus rules to v3
This upgrades to the latest format for the consensus rules. This isn’t
strictly necessary as gitconsensus has backwards compatibility support,
but it will let us add new features in the future and keeps us from
accruing technical debt.

This PR will not change anything with regards to how votes are counted
and PRs merged- it leaves all rules the same.